pagine_gialle_scraper/requirements.txt
amqp==5.1.1
arrow==1.2.3
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
billiard==3.6.4.0
blessed==1.20.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
crispy-bootstrap4==2022.1
Django==4.2.1
django-autoslug==1.9.9
django-bootstrap-v5==1.0.11
django-crispy-forms==2.0
django-filter==23.2
django-picklefield==3.1
django-q==1.3.9
idna==3.4
kombu==5.2.4
numpy==1.24.3
pandas==2.0.1
prompt-toolkit==3.0.38
python-dateutil==2.8.2
pytz==2023.3
redis==3.5.3
requests==2.30.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
